Add code to generate .webp images but do not enable it yet.
This requires updated head.html to dynamically change all the URLs to load the
webp files instead, but before that we need to ensure that the webp files are
actually smaller, which is not always the case.

Sadly the constrast+pngcrush operation doesn't seem to be deterministic so some
PNG files were modified. :(